// web/src/schema/product.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Schema for a Cannabis Product
 */
export type Product = {
  [k: string]: unknown;
} & {
  /**
   * Product ID (Autoset on Convert)
   */
  id?: string;
  /**
   * Reference to Strain ID (Autoset on Convert)
   */
  strain_id?: string;
  /**
   * Name of the product (e.g. OG Kush)
   */
  name: string;
  /**
   * Product Details
   */
  details:
    | {
        /**
         * type => Medical
         */
        type: "Medical";
        /**
         * Vendor name for medicals
         */
        vendor: string;
        [k: string]: unknown;
      }
    | {
        /**
         * type => Recreational
         */
        type: "Recreational";
        /**
         * Breeder name
         */
        breeder: string;
        /**
         * Distribution form
         */
        distribution?: "Seeds" | "Cut";
        [k: string]: unknown;
      };
};