// web/src/schema/report.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Schema for a Product Report
 */
export type Report = {
  [k: string]: unknown;
} & {
  /**
   * Report ID (Autoset on Convert)
   */
  id?: string;
  /**
   * Related Product ID (Autoset on Convert)
   */
  product_id?: string;
  /**
   * Autor of the Report
   */
  user: string;
  /**
   * Date of the Report
   */
  date: string;
  /**
   * Autors Rating from 0 to 10
   */
  rating: number;
  /**
   * Title of the Report
   */
  title: string;
  /**
   * Description of the Report
   */
  description: string;
  /**
   * List of Images for the Report
   */
  images?: string[];
  /**
   * Report Details
   */
  details?:
    | {
        /**
         * Type => Flower Rosin Report
         */
        type: "Flower Rosin";
        /**
         * Pressing BagSize in Mycron
         */
        bagsize?: number;
        /**
         * Pressing Duration in Secounds
         */
        duration?: number;
        /**
         * Pressing Temperature in C°
         */
        temperature?: number;
        /**
         * Rosin yield in %
         */
        yield: number;
        [k: string]: unknown;
      }
    | {
        /**
         * Type => Washing Report
         */
        type: "Bubble Hash";
        /**
         * Bubble Hash yield in %
         */
        yield: number;
        /**
         * Washing Bags Sizes in Mycron
         */
        bagsize?: string[];
        [k: string]: unknown;
      }
    | {
        /**
         * Type => Grow Report
         */
        type: "Grow";
        /**
         * Grow yield in Gram per Watts
         */
        yield: number;
        /**
         * Grow Weeks Seed to Harvest
         */
        weeks?: number;
        [k: string]: unknown;
      }
    | {
        /**
         * Type => Shopping Report
         */
        type: "Shop";
        /**
         * Shopping Price/Gram in €
         */
        price: number;
        [k: string]: unknown;
      };
  [k: string]: unknown;
};